(PC-34306): create newBookingPage with Ticket cutout's upperpart logic
