+++ src/features/bookings/fixtures/bookingsSnap.ts
@@ -241,6 +241,68 @@ export const bookingsSnap = toMutable({
         { barcode: 'PASSCULTURE:v3;TOKEN:352UW4', seat: 'A13' },
       ],
     },
+    {
+      id: 125,
+      cancellationDate: null,
+      cancellationReason: null,
+      confirmationDate: '2024-03-15T23:01:37.925926',
+      dateCreated: '2024-02-15T23:01:37.925926',
+      dateUsed: null,
+      expirationDate: null,
+      totalAmount: 1900,
+      token: '352UW4',
+      quantity: 100,
+      qrCodeData: 'PASSCULTURE:v3;TOKEN:352UW4',
+      canReact: false,
+      enablePopUpReaction: false,
+      stock: {
+        id: 150230,
+        beginningDatetime: '2021-03-15T20:00:00',
+        price: 100,
+        priceCategoryLabel: 'Cat 4',
+        features: [],
+        offer: {
+          id: 147874,
+          bookingContact: null,
+          name: 'Un titre de livre',
+          address: {
+            street: '45 rue de Paris',
+            postalCode: '59300',
+            city: 'Valenciennes',
+            label: null,
+            coordinates: {
+              latitude: 50.3588,
+              longitude: 3.5217,
+            },
+            timezone: 'Europe/Paris',
+          },
+          extraData: {
+            ean: '123456789',
+          },
+          isPermanent: false,
+          isDigital: false,
+          subcategoryId: SubcategoryIdEnum.LIVRE_PAPIER,
+          venue: {
+            id: 2185,
+            city: 'Valenciennes',
+            name: 'Libraire du Petit Prince',
+            coordinates: {
+              latitude: 50.3588,
+              longitude: 3.5217,
+            },
+            address: '45 rue de Paris',
+            postalCode: '59300',
+            timezone: 'Europe/Paris',
+          },
+          withdrawalDetails: 'How to withdraw, https://test.com',
+          withdrawalType: WithdrawalTypeEnum.on_site,
+        },
+      },
+      externalBookings: [
+        { barcode: 'PASSCULTURE:v3;TOKEN:352UW4', seat: 'A12' },
+        { barcode: 'PASSCULTURE:v3;TOKEN:352UW4', seat: 'A13' },
+      ],
+    },
   ],
   hasBookingsAfter18: true,
 } as const satisfies ReadonlyDeep<BookingsResponse>)
